LayoutTests/imported/w3c/web-platform-tests/css/css-animations/parsing/animation-trigger-exit-range-shorthand-expected.txt


FAIL e.style['animation-trigger-exit-range'] = "normal" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "normal normal" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "cover" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "contain" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry-crossing" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "exit" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "exit-crossing" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry, exit" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry 0% entry 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry-crossing 0% entry-crossing 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "exit 0% exit 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "exit-crossing 0% exit-crossing 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "cover 0% cover 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "contain 0% contain 100%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry calc(10% - 10%) entry calc(50% + 50%)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "cover 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "contain 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry-crossing 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "exit-crossing 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry 50px exit 100px" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "exit calc(10% + 50px)" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "entry 50% exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "cover 50% entry 50%, contain 50% exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "50% exit 50%" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "normal 100px" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "100px" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "100px normal" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL e.style['animation-trigger-exit-range'] = "10% normal" should set the property value assert_not_equals: property should be set got disallowed value ""
FAIL Property animation-trigger-exit-range value 'normal' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'normal normal' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'cover' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'contain' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry-crossing' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'exit' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'exit-crossing' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry, exit' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry 0% entry 100%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry-crossing 0% entry-crossing 100%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'exit 0% exit 100%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'exit-crossing 0% exit-crossing 100%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'cover 0% cover 100%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'contain 0% contain 100%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry calc(10% - 10%) entry calc(50% + 50%)' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'cover 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'contain 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry-crossing 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'exit 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'exit-crossing 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry 50px exit 100px' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'exit calc(10% + 50px)' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry 50% exit 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'cover 50% entry 50%, contain 50% exit 50%' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'entry 10em exit 20em' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value '10em exit 20em' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value 'normal 100px' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value '100px' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value '100px normal' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value '10% normal' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
FAIL Property animation-trigger-exit-range value '10% calc(70% + 10% * sign(100em - 1px))' assert_true: animation-trigger-exit-range doesn't seem to be supported in the computed style expected true got false
PASS e.style['animation-trigger-exit-range'] = "entry 50% 0s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "0s entry 50%" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "-1s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s 2s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s, 2s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s 2s, 3s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s, 2s 3s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s, 2s, 3s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s 2s 3s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "0s, 1s 2s 3s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s / 2s" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "1s, 2px" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "#ff0000" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "red" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "thing" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "thing 0%" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "thing 42%" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "thing 100%" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "thing 100px" should not set the property value
PASS e.style['animation-trigger-exit-range'] = "100% thing" should not set the property value
FAIL e.style['animation-trigger-exit-range'] = "normal" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "normal normal" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal normal" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal normal" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "normal entry 100%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "entry" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal entry 100%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal entry 100%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "normal entry 10%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "entry 10%" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal entry 10%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "normal entry 10%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "cover" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "cover" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "cover" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "cover" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "contain" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "contain" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "contain" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "contain" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "contain 100% contain 0%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "contain 0%" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "contain 100% contain 0%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "contain 100%" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "contain 100% contain 0%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "entry 10% exit 20%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "exit 20%" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry 10% exit 20%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "entry 10%" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry 10% exit 20%" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "entry calc(10% + 10px) exit 20px" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "exit 20px" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry calc(10% + 10px) exit 20px" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "entry calc(10% + 10px)" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry calc(10% + 10px) exit 20px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "entry, exit" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "entry, exit" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry, exit" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "entry, exit" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry, exit" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "entry 0%, exit" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "entry, exit" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry 0%, exit" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "entry, exit" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "entry 0%, exit" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "exit calc(10% + 50px)" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "exit" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "exit calc(10% + 50px)" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "exit calc(10% + 50px)" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "exit calc(10% + 50px)" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "100px" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "100px" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "100px" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "100px" should not set unrelated longhands assert_true: expected true got false
FAIL e.style['animation-trigger-exit-range'] = "10%" should set animation-trigger-exit-range-end assert_equals: animation-trigger-exit-range-end should be canonical expected (string) "normal" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "10%" should set animation-trigger-exit-range-start assert_equals: animation-trigger-exit-range-start should be canonical expected (string) "10%" but got (undefined) undefined
FAIL e.style['animation-trigger-exit-range'] = "10%" should not set unrelated longhands assert_true: expected true got false
